requirements.txt
flask
flask-smorest
flask-cors
python-dotenv
marshmallow
sqlalchemy
flask-sqlalchemy
flask-migrate
passlib
psycopg2-binary
flask-jwt-extended
pytest